pyproject.toml
[project]
name = "open-webui-knowledge-sync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "typer>=0.15.2",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "open-webui>=0.5.20",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "ruff>=0.11.2",
]

[project.scripts]
owui_sync = "open_webui_knowledge_sync.entrypoints.sync:app"
owui_watch = "open_webui_knowledge_sync.entrypoints.watch:app"

[tool.uv]
package = true
default-groups = []
